refactor: move schema validators
